+++ lib/src/televerse/bot/bot.dart
@@ -589,17 +589,19 @@ class Bot<CTX extends Context> {
       await _initializeBot();
     }
 
+    fetcher._updateStreamController = StreamController<Update>.broadcast();
     fetcher._updateSubscription = fetcher.onUpdate().listen(
           _onUpdate,
           onDone: _onStop,
         );
+
     try {
       await fetcher.start();
     } catch (err, stack) {
-      await fetcher.pause();
+      await fetcher.stop();
       final botErr = BotError<CTX>(err, stack);
       await _onError(botErr);
-      await fetcher.resume();
+      await fetcher.stop();
     }
   }
 
@@ -611,16 +613,6 @@ class Bot<CTX extends Context> {
       api.closeClient();
     }
     return fetcher.stop();
-  }
-
-  /// Pause listening of updates.
-  Future<void> pause() {
-    return fetcher.pause();
-  }
-
-  /// Resume listening of updates.
-  Future<void> resume() {
-    return fetcher.resume();
   }
 
   /// Stream of [Update] objects.